Moved tests to end of src/modules/docker.rs
